Move platform-agnostic ping senders into mzutils
